+++ sarkit/cphd/__init__.py
@@ -23,13 +23,23 @@ Data Structure & File Format
    Reader
    Writer
    read_file_header
-   get_pvp_dtype
    binary_format_string_to_dtype
    dtype_to_binary_format_string
    mask_support_array
 
+Per Vector Parameters
+
+.. autosummary::
+   :toctree: generated/
+
+   dtype_to_pvp_element
+   get_pvp_dtype
+   get_defined_pvp_dtype
+
 XML Metadata
 ============
+
+High-level classes and methods
 
 .. autosummary::
    :toctree: generated/
@@ -37,6 +47,12 @@ XML Metadata
    XmlHelper
    ElementWrapper
    XsdHelper
+
+Transcoders for defined XML parameter types
+
+.. autosummary::
+   :toctree: generated/
+
    TxtType
    EnuType
    BoolType
@@ -53,7 +69,14 @@ XML Metadata
    Poly2dType
    XyzPolyType
    PvpType
+
+Derived/convenience transcoders
+
+.. autosummary::
+   :toctree: generated/
+
    AddedPvpType
+   DefinedPvpType
    ImageAreaCornerPointsType
    ParameterType
 
@@ -156,9 +179,13 @@ from ._io import (
     Writer,
     binary_format_string_to_dtype,
     dtype_to_binary_format_string,
-    get_pvp_dtype,
     mask_support_array,
     read_file_header,
+)
+from ._pvps import (
+    dtype_to_pvp_element,
+    get_defined_pvp_dtype,
+    get_pvp_dtype,
 )
 from ._refgeom import (
     compute_reference_geometry,
@@ -181,6 +208,7 @@ from ._xml import (
     AddedPvpType,
     BoolType,
     DblType,
+    DefinedPvpType,
     ElementWrapper,
     EnuType,
     HexType,
@@ -209,6 +237,7 @@ __all__ = [
     "AddedPvpType",
     "BoolType",
     "DblType",
+    "DefinedPvpType",
     "ElementWrapper",
     "EnuType",
     "FileHeaderPart",
@@ -237,7 +266,9 @@ __all__ = [
     "compute_t_ref",
     "compute_t_ref_from_pvps",
     "dtype_to_binary_format_string",
+    "dtype_to_pvp_element",
     "ecf_to_iac",
+    "get_defined_pvp_dtype",
     "get_pvp_dtype",
     "hae_iac_to_llh",
     "hae_llh_to_iac",